Move views to print package
