+++ app/controllers/print/PrintLastDeclaredController.scala
@@ -22,7 +22,7 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.MessagesControllerComponents
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import utils.print.PrintPlaybackHelper
-import views.html.PlaybackAnswersView
+import views.html.print.PlaybackAnswersView
 
 import scala.concurrent.{ExecutionContext, Future}
