+++ app/controllers/print/PrintMaintainDeclaredAnswersController.scala
@@ -28,7 +28,7 @@ import uk.gov.hmrc.auth.core.AffinityGroup.Agent
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import utils.DateFormatter
 import utils.print.PrintPlaybackHelper
-import views.html.declaration.{PlaybackDeclaredAnswersView, PlaybackFinalDeclaredAnswersView}
+import views.html.print.{PlaybackDeclaredAnswersView, PlaybackFinalDeclaredAnswersView}
 
 import scala.concurrent.{ExecutionContext, Future}
