+++ app/controllers/print/PrintMaintainDraftAnswersController.scala
@@ -22,7 +22,7 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import utils.print.PrintPlaybackHelper
-import views.html.declaration.PlaybackDraftAnswersView
+import views.html.print.PlaybackDraftAnswersView
 
 import scala.concurrent.{ExecutionContext, Future}
